fix(audit): pass 1 — bump the wire to v9, pair the disconnect, split the wheel TU
Seven findings from an audit of #1029.

HIGH. The new required signal did not bump the wire. ApiVersion and
MinPeerApiVersion stayed at 8 and the note was filed inside the v5 block,
which is the already-shipped v3.4 wire. v8 shipped in 3.4.4, and the
ledger's own v8 entry sets the precedent verbatim: a signal-only addition
is not harmlessly additive, because a mismatched pair matches every
signature, errors nowhere, and then goes on scrolling and parking a column
whose window the compositor is refusing to move. Both constants go to 9,
the note becomes its own v9 entry, and the pinning test in
test_phosphorprotocol.cpp follows.

MEDIUM. connectSignals' new bus.connect had no bus.disconnect above it,
which is what that block exists for: onDaemonReady re-runs it on every
daemon restart, and Qt registers the same handler again. Paired.

MEDIUM. CHANGELOG had nothing under Unreleased. Adds the Fixed entry for
the keyboard half and the Changed entry for the wire break, the one 3.4.5
carried for the same reason. The 3.4.5 text is left alone: it is an
accurate record of what 3.4.5 did.

MEDIUM. state.cpp went 1179 -> 1203, past the 1150 ceiling it was already
over. The wheel-chord dispatch is split into wheelchord.cpp: the chord,
its accumulators, the cursor-resolved target screen, the fullscreen
release and the slot that carries the daemon's request. None of it is the
daemon-published per-session state state.cpp is about. state.cpp is now
823 lines and the moved code is byte-identical bar the one comment below.

LOW. The reapplyWindowGeometry comment claimed both its gates came from
wheelTargetScreen, which is only true of the wheel caller. Rewritten to
say how each caller satisfies them.

LOW. leaveNativeFullscreenTiles was public while its own doc named two
in-class callers. Moved to private beside wheelTargetScreen.

LOW. The Scrolling interface DocString enumerates the interface and had
not gained the new signal.

Build clean with no new warnings, 417/417 tests pass. The three touched
effect TUs were also compiled standalone with unity off, since the default
build batches them.
