+++ dbus/org.plasmazones.Scrolling.xml
@@ -6,7 +6,7 @@ SPDX-License-Identifier: GPL-3.0-or-later
 -->
 <node>
     <interface name="org.plasmazones.Scrolling">
-        <annotation name="org.gtk.GDBus.DocString" value="Scroll-engine-specific surface. Window lifecycle and tile-request traffic for scrolling screens rides org.plasmazones.Tiling (the shared tiling-family pipeline, which routes per screen). This interface carries the scrolling screen set, the strip snapshot the settings app renders as its preview surface (plus the preset vocabulary beside it), the wheel-driven focusColumn and scrollView verbs, the four absolute width/height setters for external scripting, the toggleMaximizeColumn width verb, the toggleMaximizeToEdges verb the KWin effect sends when a scroll-managed window is asked to maximize, the clearWindowedFullscreen reconciliation call the KWin effect sends when a client leaves fullscreen on its own, the reapplyWindowGeometry repair call it sends on a fullscreen exit whose strip rects never moved, the blueprintProgressJson template-seed report, the stripChanged wake-up that tells a renderer its strip is worth re-reading, the focus-follows-mouse scroll cap's blocked-window list, the stripContextChanged announcement of which strip a screen is currently showing, and the scrollingScreensChanged / scrollEffectBehaviourChanged / scrollFocusScrollBlockedWindowsChanged change announcements for the three properties."/>
+        <annotation name="org.gtk.GDBus.DocString" value="Scroll-engine-specific surface. Window lifecycle and tile-request traffic for scrolling screens rides org.plasmazones.Tiling (the shared tiling-family pipeline, which routes per screen). This interface carries the scrolling screen set, the strip snapshot the settings app renders as its preview surface (plus the preset vocabulary beside it), the wheel-driven focusColumn and scrollView verbs, the four absolute width/height setters for external scripting, the toggleMaximizeColumn width verb, the toggleMaximizeToEdges verb the KWin effect sends when a scroll-managed window is asked to maximize, the clearWindowedFullscreen reconciliation call the KWin effect sends when a client leaves fullscreen on its own, the reapplyWindowGeometry repair call it sends on a fullscreen exit whose strip rects never moved, the blueprintProgressJson template-seed report, the stripChanged wake-up that tells a renderer its strip is worth re-reading, the focus-follows-mouse scroll cap's blocked-window list, the stripContextChanged announcement of which strip a screen is currently showing, the leaveNativeFullscreenRequested release the daemon sends ahead of a keyboard strip verb, and the scrollingScreensChanged / scrollEffectBehaviourChanged / scrollFocusScrollBlockedWindowsChanged change announcements for the three properties."/>
 
         <property name="scrollingScreens" type="as" access="read">
             <annotation name="org.gtk.GDBus.DocString" value="List of screen IDs currently using the scrolling engine, sorted, matching the ordering of the scrollingScreensChanged payload so a consumer can compare the two for the same set. The KWin effect uses it as the discriminator that stamps Mode 'scrolling' rather than 'tiling' in window rule queries, but it does not answer from this raw set: it intersects it at read time with the engine-managed union published on org.plasmazones.Tiling, because the two sets arrive as independent signals with no ordering guarantee and this one can transiently name a screen the union has already dropped. Changes are announced on the scrollingScreensChanged signal below, not through org.freedesktop.DBus.Properties.PropertiesChanged, which this property does not emit."/>
