+++ kwin-effect/tilinghandler/tilinghandler.h
@@ -579,19 +579,6 @@ public:
     bool handleWheelChord(qreal delta, qint32 deltaV120, Qt::Orientation orientation, Qt::KeyboardModifiers mods,
                           Qt::MouseButtons buttons);
 
-    /// Leaves the OWN fullscreen (a client F11, a video going fullscreen — NOT
-    /// the windowed-fullscreen feature) of every scroll-tracked tile on
-    /// @p screenId. A tile in that state refuses every geometry commit through
-    /// applyWindowGeometry's fullscreen bail while the engine goes on scrolling
-    /// and PARKING its column, so the two owners drift apart for the whole hold.
-    ///
-    /// Called from the two USER-VERB dispatch sites and nowhere else: the wheel
-    /// chord above, and the daemon's keyboard shortcut gate arriving over
-    /// Scrolling.leaveNativeFullscreenRequested. Both call it BEFORE their verb
-    /// goes out. Not callable from the batch apply — see the site comment for
-    /// the measurement that rules that out.
-    void leaveNativeFullscreenTiles(const QString& screenId);
-
     // Screen accessors (for gating drag/snap/overlay behavior per-screen)
     bool isManagedScreen(const QString& screenId) const;
 
@@ -1509,6 +1496,20 @@ private:
     /// gesture. Scrolling over a strip on the second monitor should move that
     /// strip, not the one holding focus.
     QString wheelTargetScreen() const;
+
+    /// Leaves the OWN fullscreen (a client F11, a video going fullscreen — NOT
+    /// the windowed-fullscreen feature) of every scroll-tracked tile on
+    /// @p screenId. A tile in that state refuses every geometry commit through
+    /// applyWindowGeometry's fullscreen bail while the engine goes on scrolling
+    /// and PARKING its column, so the two owners drift apart for the whole hold.
+    ///
+    /// PRIVATE on purpose. It has exactly two callers, both in-class and both
+    /// USER-VERB dispatch sites: handleWheelChord, and
+    /// slotLeaveNativeFullscreenRequested carrying the daemon's keyboard
+    /// shortcut gate over Scrolling.leaveNativeFullscreenRequested. Both call it
+    /// BEFORE their verb goes out. Not callable from the batch apply — see the
+    /// site comment in wheelchord.cpp for the measurement that rules that out.
+    void leaveNativeFullscreenTiles(const QString& screenId);
 
     /// Drop any banked sub-notch remainder. Called from every path that stops
     /// claiming axis events, so a partial notch cannot outlive the gesture
